WIP -Add device tests for PWM and Tone test/dev
This is a WIP and includes a PowerShell script (yes, I hear the groans
already), but it will be moved to Python soon to allow cross-platform
compatibility.

A new directory under tests is added (because the tests/device are
self-contained and doesn't easily support the needs of this test).

A USB logic analyzer is required to be hooked up to the 8266, and a test
slave script should be uploaded.

The main test script, run on the host, will reset the chip, then run a
series of tests varying some parameter or other and collect individual
waveforms at each test point.

These raw waveforms are the post-processed to determine exact, per-cycle
phase, period, etc. parameters at that test point.

This will be repeated for every test point desired (the checked in
version now starts 2 25MHz analogWrites, one moviing from 0...1000 and
one moving from 1000...0) and calculates the performance.

Eventually this will be part of a makefile with a series of test scripts
running on a more general-purpose host test driver.