Make Python classes for SUT and LA, check results
Use a simpler setup where it is possible to write a Python script just
like you'd write an INO on the 8266:
    sut.reset()
    sut.analogWriteFreq(10000)
    sut.analogWrite(2, 300)
    ....

Also, check the output of the SUT after each command.  If there is an
error, throw an exception and abort the test.

Add CPU speed in MHz and the ESP.getVersion() string to the test to help
identify the source of the code being tested.