Fix 1st cycle measurement, DC reporting
If the waveform started with a 1, the initial cycle would end up being
garbage.  Fix this by looking for a 0 before looking for a 1.

Report DC signals better (duty cycle, etc.) to allow graphing w/o
futzing.  The period for DC is still "infinite" since there is no
information as to what it's supposed to be.