Change docs to use a GH Deploy Key
The original comments suggested adding a new SSH key to a user account
which would have allowed complete access to all repos and all other
global user settings.  A really bad thing if it got into the wild.

Instead, a single deploy key can be added to a single repo.  This only
grants git push access/etc. for that single repo and not for the entire
user account.  Update the docs accordingly.

Also rename push.yml->pull-request.yml since it runs on a pull request,
not a master push.