Use root cert, not fingerprint for api.github.com
The HTTPS example we were using a fingerprint which seems to change
daily as the github.com certificates are regenerated.  Replace this with
a trust anchor based on the ultimate root CA that github.com uses to
sign their certificates.  Assuming they don't change CAs, this
certificate should be good until 2030+

Fixes #7489