Handle breakpoints via postmortem when GDB is not running.
In the absence of GDB, install a small breakpoint handler that clears EXCM and
executes an illegal instruction. This leverages the existing exception
handling code in the SDK for `ill` instruction. At postmortem processing,
the real event is identified and reported. This allows for alerting of embedded
breakpoints. In the stack trace, `epc1` is updated with the value of `epc2` so
that the "ESP Exception Decoder" will identify the line where the BP occurred.

The SDKs default behavior was to loop on breakpoints waiting for interrupts
>2 until the HWDT kicks in. The current "C++" compiler will embed breakpoints
when it detects a divide by zero at compile time.

Expanded technique to preserve more of the current state and to also report
Kernel and Double exceptions through postmortem.