fix by @devyte's internal review:
At this point it was found that wpos is above rpos, i.e.: the ringbuffer has
a single contiguous block, with possible free mem at the bottom and at the
top.
If at this point there is an interrupt that causes wpos to wrap, the
returned value below will be wrong, i.e.: some large number because of
negative result returned as size_t type.  Then, in SendGenericPeekBuffer(),
the to->write() call will crash.