Add errors on invalid/missing function return type (#8165)
GCC 10.x seems to have a knack for crashing when a function which is declared
to return a value does not.  Add a warning, present on all builds, when this
is the case.  For more info see https://github.com/esp8266/Arduino/discussions/8160

Thanks to @hreintke and @mcspr for the tips.