Integrated handling for filesystem and gzipped Binaries
ATENTION: This is tested on Linux only! See below for open questions/prerequisites.

Currently there are some manual steps necessary to prepare the files necessary for OTA updates (especially for ESP8266HTTPUpdateServer)

The steps are:
1. Compile the sketch
2. Export the generated sketch-binary
3. Generate the filesystem
4. Export the fs-binary
5. Generate gzipped versions of binaries
6. Generate signed versions of gzipped binaries

Which of the above steps are necessary depends on the sketch (does it use a fs?) and the kind of OTA chosen (signed binaries necessary?) or preferred usage (use of gzipped binaries for faster upload?).

There are some problems/inconveniences:
2. The signed sketch-binary is not exported when using "Export compiled Binary".
(This point will be solved with my Pull-Request #8255)
3. & 4. AFAIK, the only “integrated” way to generate the fs is using one of the tools:
“arduino-esp8266littlefs-plugin” or  “arduino-esp8266fs-plugin”.
Both tools are designed to create & upload the fs, but not export them.
So if preparing the files for an OTA when no board is connected, you have to:
- ignore errors shown because upload is not possible
- find & navigate to the “build.path” and copy the fs-binary manually
5. & 6. AFAIK, there is no “integrated” way to perform these steps.

This Pull-Request will integrate all of the above steps into the “compile and/or upload” action of Arduino.
This is done by adding two new entries (“Filesystem” and “Binaries”) to the tools-menu.
These entries are presented just like any other options for esp8266 (“Upload speed” ... “Non-32-Bit Access”) and placed right after the last.

Menu-Entry “Filesystem” let You choose between:
- Off
- LitteFs: Create & Upload
- LitteFs: Create only
- SPIFFS: Create & Upload
- SPIFFS: Create only

Menu-Entry “Binaries” let You choose between:
- Off
- Export (.bin & .bin.signed)
- Create & Export gzipped Binaries too

Now You can choose with of the steps mentioned above will be automatically performed.
After a (successful) compile, the files will be exported to a subdir “bin/{variant}” of Your sketch directory (“{variant}” replaced with the name of the board as shown behind “Tools -> Board:”).
Signed variants of all Binaries are gererated similar to the “automatic signing” done for normal sketch binary.

Open questions/prerequisites:
As stated above I had tested this on Linux only.
Even under this os there are some prerequisites:
- gzip must be available (inside path)
if it is not, a message is given and creating of gzipped binaries is skipped.
- tkinter should be available
if it is not, and creating of fs is requested AND the data dir is empty, I didn't know how to ask the user: “Are you sure you want to create an empty ... image?”.
So in this case a message is given and the fs will not be created.

This “should” work under Windows/MacOs the same, but can't be tested by me (I had no PC with this OSes)

An additional question for developers:
As mentioned inside function main of “postbuild.py”, there is a strange behavior of the arg "{build.path}" supplied via “platform.txt”. If someone can explain me why this “string” behave as described in the comments, I would be happy ;-)