Create Range_Extender_with_WiFi_Manager.ino
WiFi range extender with a WiFi manager. 
Allows you to change the WiFi network the extender is connected to
Led blinks at different rates to indicate various states of the extender (useful for debugging)
GPIO 0 is used as a reset button: this deletes the WiFi credentials stored in the file system.