API in WiFiClient to save heap memory
When used by a server heap memory decrease drastically after numero client acces. Sockets in TIME_WAIT stay after the server issues a Client.stop which causes a server crash (and reboot) after some hours or some minutes. Note that the proposed method tcpCleanup in #1923 does not work because the instruction "extern struct tcp_pcb* tcp_tw_pcbs;" does NOT initialise tcp_tw_pcbs to the one used by the Client. It is only a declaration and because it is only a declaration, tcp_tw_pcbs is set to NULL. Consecuently tcp_abort(tcp_tw_pcbs) is never called.